Minor Auth Refactor
