+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-import { handlers } from "@/auth";
+import { handlers } from "@/authentication/auth";
 
 export { handlers as GET, handlers as POST };
